Move all to sync
